Virgin import of patch-2.5.9, the last official GPLv2 release.
As in previous imports we dropped the mkinstall and pc directories
as they are of no use to us.

Differently from previous imports, patch.1 was not renamed and
is maintained with its original name (patch.man). If we ever
merge this into the main tree, care must be taken to rename
it back.