Import DRM core files and compat code
Import DRM code files from Linux 5.10
For arm/arm64 single board computer we cannot import the Linux ones as
they are GPLv2 licenced so we need to make native FreeBSD drivers.
Therefore we need some compat code not compatible with Linuxkpi to be
present (as the drivers aren't linux driver but native one).
This is the goal of drmcompat directory and kernel options to turn this
part of the code on or not.

Co-authored-by:	Michal Meloun <mmel@FreeBSD.org>
Co-authored-by:	Ruslan Bukin <br@FreeBSD.org>
Co-authored-by:	Vladimir Kondratyev <wulf@FreeBSD.org>

Sponsored by:	Beckhoff Automation GmbH & Co. KG