drm: Add rockchip VOP driver
This is the display controller driver present on some RockChip SoC.
It provides a way to transfer image data from a video memory buffer
to a external LCD interface or bridge.
The IP supports two VOP on some SoCs to provides multiple outputs
at the same time.
As off today only HDMI endpoint is supported and only VOP-little
is supported too.