pkgbase: Always stage the kernel using INSTKERNNAME
When building with only one kernel stage it and create the package
so it will be installed in /boot/kernel.${INSTKERNNAME}.
The advantage is that is doesn't conflict with make installkernel which helps
when you are using pkgbase on your main dev box.

Sponsored by:	Beckhoff Automation GmbH & Co. KG